Save function is complete
Finished the save function, it works, as well as i have a format for identifying data for the load function which is next on the list.
at the same time, origin print(debugging function) recieved some corrections to fix bugs in which nodes were printed(some were omitted). this made the save function difficult to debug, but these have been corrected, as well as minor changes in the blank map creation and the random map generation.